Speed up stack decoding, optimise speedscope output size and more sample helpers
Big commit with lots of changes!

This limits the number of stack frames to 255 to make sure the
eventlog message doesn't exceed 2^16 bytes. Naturally, this may lead to
cut off stack traces.

During empirical experiments with GHC, we noticed a memory leak and
spending excessive amount of time on looking up `ProvInfo` only to
ignore most of it and only using the `ProvInfoId`. We no longer use
`lookupIPE` during stack decoding, speeding up the sampling process.

Further, we speed up generation of the speedscope json output, by only
including IPE entries that are actually used by the profile.
Previously, we included all IPE entries of the program.
This dramatically reduces the file size of the profile, at the cost of
slightly increased memory usage of the generation program.

We also expand the API for instrumenting specific threads for profile
samples.
Users can decide to instrument all alive threads, only a specific thread
or the current thread only.