Allow arbitrary depth callstacks
Huge change, let's break it down. The goal of this commit is to enable
`ghc-stack-profiler` to send callstacks of arbitrary depth to the
eventlog. The protocol we came up with can't handle this right now. So,
we extend the protocol to send multiple callstack messages with markers
to indicate the given callstack message is not the final one.

Further, we want to be able to trivially tell the size of a
'BinaryCallStackMessage' without having to look at each individual
message to make it easier to chunk it into multiple messages.
So, we replace arbitrary size data messages, namely 'Text' and
'SourceLocation's with IDs and send them as separate messages.
This allows us to reduce bandwidth, at the cost of storing additional
state in the profiler (namely a Map from these symbols to IDs).

With these two extensions in mind, the eventlog protocol messages now
look like this:

    MESSAGE
     := "FF" "CA" <stack: STACK>
      | "FF" "CB" <prefix: STACK>
      | "FF" "CC" <stringId: Word64> <string: CStringLen>
      | "FF" "CD" <srcLocId: Word64> <row: Word32> <col: Word32> <functionId: Word64> <filename: Word64>

    STACK
     := <capability: Word32> <threadId: Word32> <length: Int16> <ENTRY>+
      # check that length < (2^16-8) / 9

    ENTRY
     := "01" <ipe: Word64>
      | "02" <stringId: Word64>
      | "03" <srcLocId: Word64>

    CStringLen
      := <length: Int16> <Char>+
       # check that length < 2^16-8

These changes to the protocol require us to re-engineer large parts of
the codebase.
Most of these are logical consequences, but a few unrelated changes
and documentation have snuck in as well.

To implement the dehydration in the profiler, we introduce the concept
of a 'SymbolTableWriter'. It is an abstract interface to deduplicate the
symbols we care about. The implementation can be changed on a per-need
basis.
Similarly, we introduce 'SymbolTableReader' for the speedscope
deserialiser.