Add compatibility for GHC 9.14 and 9.12
Backports changes and the RTS callstack decoder for use with GHC 9.14
and 9.12.

The backports are only included on older versions of GHC.
Further, the Compat modules are written in a forward compatible way,
such that dropping them should be easy in the future.

Also, since we finally support standard GHC versions, let's add CI!