Implement support for eventlog-socket and custom commands
We want to be able to stop and start the profiler from the outside
world.
The first thing we do, is to create a central object that can stop the
profiler threads and let them wait until it is started again.
We call this the `StackProfilerManager` object, of which there should be
currently only a single object.
In the future, we would like to add support for `ki` similar scoping
mechanisms. Right now, the root `StackProfilerManager` needs to be
manually initialised and then can be used to start profiler threads.
This is quite manual, so perhaps we want to store the
`StackProfilerManager` as a global object, and provide variants that
implicitly use the root `StackProfilerManager` without needing to
explicitly passing the root profiler object.

The user can create signal handlers or other logic to stop and start the
profiler threads using the `StackProfilerManager` object.
Further, we implement built-in support for `eventlog-socket`.
We install custom handlers that allow users to control the profiler via
`eventlog-socket`, which will in the long run allow us to start and stop
the profiler from grafana or similar services.